requirements.txt
attrs==20.3.0
boto3==1.17.55
botocore==1.20.55
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==8.1.3
cryptography==37.0.2
dnspython==2.1.0
Flask==2.1.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.22.4
oauthlib==3.1.0
packaging==20.9
pandas==1.4.2
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.10.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.1
python-dateutil==2.8.1
pytz==2022.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.22
threadpoolctl==3.1.0
toml==0.10.2
tweepy==3.10.0
urllib3==1.26.2
Werkzeug==2.1.2
xgboost==0.90
zipp==3.8.0
